dom/base/test/test_bug869006.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=869006
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 869006</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=869006">Mozilla Bug 869006</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 869006 **/

var c = new Comment();
ok(c, "Comment has been created without content");
is(c.data, "", "Comment.data is ok");

c = new Comment('foo');
ok(c, "Comment has been created");
is(c.data, "foo", "Comment.data is ok");

document.appendChild(c);
ok(true, "Comment has been added to the document");

</script>
</pre>
</body>
</html>
